requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apify_fingerprint_datapoints==0.0.2
arrow==1.3.0
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
binaryornot==0.4.4
blessed==1.21.0
Brotli==1.1.0
browserforge==1.2.3
cachetools==6.0.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
cookiecutter==2.6.0
crawlee==0.6.9
cryptography==45.0.3
cssselect==1.3.0
curl_cffi==0.11.1
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
docstring_parser==0.16
editor==1.6.6
elevenlabs==2.1.0
eval_type_backport==0.2.2
fal_client==0.7.0
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
incremental==24.7.2
inquirer==3.4.0
itemadapter==0.11.0
itemloaders==1.3.2
jaro-winkler==2.0.3
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonref==1.1.0
kiwisolver==1.4.8
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.4.4
natsort==8.4.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parsel==1.10.0
pathlib==1.0.1
pathspec==0.12.1
pillow==11.2.1
playwright==1.52.0
proglog==0.1.12
propcache==0.3.1
Protego==0.4.0
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.6.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyee==13.0.0
Pygments==2.19.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
readchar==4.2.1
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
runs==1.2.2
scikit-learn==1.6.1
scipy==1.15.3
Scrapy==2.13.1
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
sievedata==1.5.0
six==1.17.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.7
sseclient==0.0.27
starlette==0.46.2
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tldextract==5.3.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
Twisted==24.11.0
typeguard==4.4.2
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.146
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.2
w3lib==2.3.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
xmod==1.8.1
yarl==1.20.0
zope.interface==7.2
zstandard==0.23.0
